requirements.txt
python-dotenv
requests
tenacity
pydantic
